Move hdmi_cec base entity to separate module
