+++ homeassistant/components/hdmi_cec/const.py
@@ -1,0 +1,7 @@
+"""Support for HDMI CEC."""
+
+DOMAIN = "hdmi_cec"
+
+ATTR_NEW = "new"
+
+EVENT_HDMI_CEC_UNAVAILABLE = "hdmi_cec_unavailable"
