+++ homeassistant/components/hdmi_cec/switch.py
@@ -12,7 +12,8 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import ATTR_NEW, DOMAIN, CecEntity
+from .const import ATTR_NEW, DOMAIN
+from .entity import CecEntity
 
 _LOGGER = logging.getLogger(__name__)
